requirements.txt
# Core Django Requirements
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Utilities
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0

# Deployment
Pillow==10.1.0